Split the visual tests into a generic runner and per-editor helpers
The runner, the way a page is driven and the two suites don't know about
any particular editor anymore: an editor is described by a helper in
`helpers/`, which is the only place knowing about it. It says how to find
its controls (by registering a resolver, so its targets are used like the
generic `{ button }`, `{ selector }`, `{ menuItem }`, `{ tab }` ones), what
its manipulations are, and how to check that what it displays matches the
project.

- `lib/PageHelpers.js` and `lib/PageDriver.js`: find and use the controls of
  any editor, including real drag and drops. `scrollList` scrolls the list
  of the page, whichever it is.
- `lib/Runner.js`: runs scripted or random manipulations and checks them,
  using the helper it is given.
- `helpers/SpriteEditor.js`: everything about the Sprite editor, moved out
  of `lib`.
- `helpers/ObjectsList.js` and `helpers/BehaviorsEditor.js`: smaller helpers
  for other parts of the app.

Add a test using them, on a completely different part of the app: it opens
an example game, searches "Fire bullet" in the behavior store, adds the
"Fire bullets" behavior to an object (its extension is downloaded and
installed) and checks it is there.

Also:
- Run the tests in the Chrome that `npm install` downloads for Puppeteer,
  instead of expecting one on the machine: the CI image has none, which was
  failing with `google-chrome: command not found`. A Chrome installed on the
  machine or given with `--chrome-path` is still used when there is one.
- Don't ask the monkey for a move that leaves the frames identical (they can
  all be the same image), as nothing could then be checked.
- Make the README much shorter.
