+++ newIDE/visual-tests/lib/Runner.js
@@ -1,17 +1,19 @@
 // @ts-check
 
 /**
- * Runs the steps of a test, or a random sequence of them (the "monkey"), and
- * checks after every single manipulation that:
- * - the page did not throw (a dangling C++ wrapper takes the whole editor down
- *   with a "memory access out of bounds" error),
- * - the editor is still there,
- * - what the editor displays is what the edited object contains (when the page
- *   exposes the animations of the object),
+ * Runs the manipulations of a test - scripted, or a random sequence of them
+ * (the "monkey") - and checks after every single one that:
+ * - the page did not throw (an uncaught error takes the whole editor down),
+ * - the editor still shows what the edited object contains (as far as the
+ *   helper of that editor can tell),
  * - the manipulation did what it was supposed to do.
+ *
+ * Everything specific to an editor comes from its helper (see
+ * `helpers/SpriteEditor.js`): the manipulations, how to describe what is
+ * displayed, and how to check it.
  */
 
-const { actions, describe, click, wait } = require('./SpriteEditorActions');
+const { closeAnyOverlay } = require('./PageDriver');
 
 /** A reproducible pseudo random number generator. */
 const makeRandom = seed => {
@@ -22,16 +24,46 @@ const makeRandom = seed => {
   };
 };
 
-const framesOf = (snapshot, position) => {
-  const animation = snapshot.animations[position];
-  if (!animation) return [];
-  return animation.frames;
-};
-
 const summarizeAnimation = animation =>
   JSON.stringify({ name: animation.name, directions: animation.directions });
 
-/** A short description of what a manipulation changed in the object. */
+/**
+ * What the edited object contains when the page gives it, or what the editor
+ * displays otherwise - so that a manipulation that did nothing can be told
+ * apart from one that worked in both cases.
+ */
+const takeSnapshot = async (page, helper) => {
+  const editedObject = helper.readEditedObject
+    ? await helper.readEditedObject(page)
+    : null;
+  if (editedObject)
+    return {
+      isFromTheObject: true,
+      animations: editedObject.map(animation => ({
+        name: animation.name,
+        frames: animation.directions[0] ? animation.directions[0].frames : [],
+        directions: animation.directions,
+      })),
+    };
+
+  const described = await helper.describe(page);
+  return {
+    isFromTheObject: false,
+    animations: (described.rows || []).map(row => ({
+      name: row.name,
+      frames: row.frames.map(frame => frame.title),
+      directions: [
+        {
+          frames: row.frames.map(frame => frame.title),
+          timeBetweenFrames: Number(row.timeBetweenFrames),
+          isLooping: row.isLooping,
+        },
+      ],
+    })),
+  };
+};
+
+/** A short description of what a manipulation changed. */
 const describeEffect = (snapshotBefore, snapshotAfter) => {
   const before = snapshotBefore.animations;
   const after = snapshotAfter.animations;
@@ -57,56 +89,92 @@ const describeEffect = (snapshotBefore, snapshotAfter) => {
   return changes.length ? changes.join(', ') : 'nothing changed';
 };
 
-/**
- * The animations of the edited object when the page exposes them (Storybook),
- * or what the editor displays otherwise (real editor) - so that a manipulation
- * that did nothing can be told apart from one that worked in both cases.
- */
-const takeSnapshot = async page => {
-  const model = await page.evaluate(() =>
-    window.gdVisualTests.readAnimations()
-  );
-  if (model)
-    return {
-      isFromTheObject: true,
-      animations: model.map(animation => ({
-        name: animation.name,
-        frames: animation.directions[0] ? animation.directions[0].frames : [],
-        directions: animation.directions,
-      })),
-    };
-
-  const described = await describe(page);
-  return {
-    isFromTheObject: false,
-    animations: described.rows.map(row => ({
-      name: row.name,
-      frames: row.frames.map(frame => frame.title),
-      directions: [
-        {
-          frames: row.frames.map(frame => frame.title),
-          timeBetweenFrames: Number(row.timeBetweenFrames),
-          isLooping: row.isLooping,
-        },
-      ],
-    })),
-  };
+const framesOf = (snapshot, position) => {
+  const animation = snapshot.animations[position];
+  return animation ? animation.frames : [];
 };
 
-const check = page => page.evaluate(() => window.gdVisualTests.check());
+/** Check the precise outcome a manipulation declared. */
+const checkExpectation = (expectation, snapshotAfter) => {
+  const problems = [];
+  const framesAfter = framesOf(snapshotAfter, expectation.row);
+  const expected = expectation.frames;
+  if (expected && framesAfter.join('|') !== expected.join('|'))
+    problems.push(
+      `the frames of the animation #${expectation.row} are [${framesAfter.join(
+        ', '
+      )}] but [${expected.join(', ')}] was expected`
+    );
+
+  const prefix = expectation.framesStartWith;
+  if (
+    prefix &&
+    framesAfter.slice(0, prefix.length).join('|') !== prefix.join('|')
+  )
+    problems.push(
+      `the frames of the animation #${expectation.row} are [${framesAfter.join(
+        ', '
+      )}] but they should start with [${prefix.join(', ')}]`
+    );
+
+  if (
+    expectation.framesCount !== undefined &&
+    framesAfter.length !== expectation.framesCount
+  )
+    problems.push(
+      `the animation #${expectation.row} has ${
+        framesAfter.length
+      } frames but ` + `${expectation.framesCount} were expected`
+    );
+
+  const animationAfter = snapshotAfter.animations[expectation.row];
+  const directionAfter = animationAfter && animationAfter.directions[0];
+  if (
+    directionAfter &&
+    expectation.timeBetweenFrames !== undefined &&
+    Math.abs(directionAfter.timeBetweenFrames - expectation.timeBetweenFrames) >
+      0.0001
+  )
+    problems.push(
+      `the animation #${expectation.row} now has ` +
+        `${directionAfter.timeBetweenFrames}s between frames instead of ` +
+        `${expectation.timeBetweenFrames}s`
+    );
+  if (
+    directionAfter &&
+    expectation.isLooping !== undefined &&
+    expectation.isLooping !== null &&
+    directionAfter.isLooping !== expectation.isLooping
+  )
+    problems.push(
+      `the looping of the animation #${expectation.row} became ` +
+        `${String(directionAfter.isLooping)}`
+    );
+  return problems;
+};
 
 /** Run one manipulation and check everything afterwards. */
-const runStep = async ({ page, pageErrors, actionName, args, random }) => {
-  const action = actions[actionName];
-  if (!action) throw new Error(`Unknown action: ${actionName}`);
+const runStep = async ({
+  page,
+  pageErrors,
+  helper,
+  actionName,
+  args,
+  random,
+}) => {
+  const action = helper.actions[actionName];
+  if (!action)
+    throw new Error(
+      `Unknown manipulation "${actionName}" for the ${helper.name} helper.`
+    );
 
-  const stateBefore = await describe(page);
+  const stateBefore = await helper.describe(page);
   const stepArgs =
     args || (action.pick ? action.pick(stateBefore, random) : {});
   if (!stepArgs) return { skipped: 'not applicable' };
 
   const description = action.describe(stepArgs);
-  const snapshotBefore = await takeSnapshot(page);
+  const snapshotBefore = await takeSnapshot(page, helper);
   const expectation = action.expect
     ? action.expect(stateBefore, stepArgs)
     : null;
@@ -123,94 +191,33 @@ const runStep = async ({ page, pageErrors, actionName, args, random }) => {
   if (pageErrors.length) return { description, crashed: true };
   if (problem) return { description, skipped: problem };
 
-  const { problems } = await check(page);
-  const snapshotAfter = await takeSnapshot(page);
+  const { problems } = await helper.check(page);
+  const snapshotAfter = await takeSnapshot(page, helper);
   const effect = describeEffect(snapshotBefore, snapshotAfter);
   const hadNoEffect = effect === 'nothing changed';
 
-  if (action.mustChangeTheObject && hadNoEffect) {
+  if (action.mustChangeTheObject && hadNoEffect)
     problems.push('the object was not changed');
-  }
 
-  if (expectation && snapshotAfter.isFromTheObject) {
-    const framesAfter = framesOf(snapshotAfter, expectation.row);
-    const expected = expectation.frames;
-    if (expected && framesAfter.join('|') !== expected.join('|')) {
-      problems.push(
-        `the frames of the animation #${
-          expectation.row
-        } are [${framesAfter.join(', ')}] but [${expected.join(
-          ', '
-        )}] was expected`
-      );
-    }
-    const prefix = expectation.framesStartWith;
-    if (
-      prefix &&
-      framesAfter.slice(0, prefix.length).join('|') !== prefix.join('|')
-    ) {
-      problems.push(
-        `the frames of the animation #${
-          expectation.row
-        } are [${framesAfter.join(
-          ', '
-        )}] but they should start with [${prefix.join(', ')}]`
-      );
-    }
-    if (
-      expectation.framesCount !== undefined &&
-      framesAfter.length !== expectation.framesCount
-    ) {
-      problems.push(
-        `the animation #${expectation.row} has ${framesAfter.length} frames ` +
-          `but ${expectation.framesCount} were expected`
-      );
-    }
-    const animationAfter = snapshotAfter.animations[expectation.row];
-    const directionAfter = animationAfter && animationAfter.directions[0];
-    if (
-      directionAfter &&
-      expectation.timeBetweenFrames !== undefined &&
-      Math.abs(
-        directionAfter.timeBetweenFrames - expectation.timeBetweenFrames
-      ) > 0.0001
-    ) {
-      problems.push(
-        `the animation #${expectation.row} now has ` +
-          `${directionAfter.timeBetweenFrames}s between frames instead of ` +
-          `${expectation.timeBetweenFrames}s`
-      );
-    }
-    if (
-      directionAfter &&
-      expectation.isLooping !== undefined &&
-      expectation.isLooping !== null &&
-      directionAfter.isLooping !== expectation.isLooping
-    ) {
-      problems.push(
-        `the looping of the animation #${expectation.row} became ` +
-          `${String(directionAfter.isLooping)}`
-      );
-    }
-  }
+  if (expectation && snapshotAfter.isFromTheObject)
+    problems.push(...checkExpectation(expectation, snapshotAfter));
 
   // A change of the animations must not leave frames selected: the selection
   // designates them by their index, which would point at other frames.
   if (action.clearsTheFrameSelection && !hadNoEffect) {
-    const stateAfter = await describe(page);
-    const stillSelected = stateAfter.rows
+    const stateAfter = await helper.describe(page);
+    const stillSelected = (stateAfter.rows || [])
       .map(row =>
         row.frames
           .filter(frame => frame.selected)
           .map(frame => `#${row.index} ${frame.title}`)
       )
       .reduce((all, some) => all.concat(some), []);
-    if (stillSelected.length) {
+    if (stillSelected.length)
       problems.push(
         `${stillSelected.length} frame(s) are still shown as selected after ` +
           `the animations changed: ${stillSelected.join(', ')}`
       );
-    }
   }
 
   if (action.keepsTheFrames) {
@@ -220,9 +227,8 @@ const runStep = async ({ page, pageErrors, actionName, args, random }) => {
         .reduce((all, frames) => all.concat(frames), [])
         .sort()
         .join('|');
-    if (allFrames(snapshotBefore) !== allFrames(snapshotAfter)) {
+    if (allFrames(snapshotBefore) !== allFrames(snapshotAfter))
       problems.push('the frames of the object were not only reordered');
-    }
   }
 
   return {
@@ -232,27 +238,6 @@ const runStep = async ({ page, pageErrors, actionName, args, random }) => {
     hadNoEffect,
     isDrop: !!action.mayHaveNoEffect,
   };
-};
-
-/** Close whatever dialog or menu is opened, so that the list is usable again. */
-const closeAnyOverlay = async page => {
-  const state = await describe(page);
-  if (state.openMenuItems.length) {
-    await page.keyboard.press('Escape');
-    await wait(300);
-    return true;
-  }
-  if (state.openDialogTitles.length) {
-    const closed =
-      (await click(page, { global: 'exactButton', text: 'Apply' })) ||
-      (await click(page, { global: 'exactButton', text: 'Ok' })) ||
-      (await click(page, { global: 'exactButton', text: 'Close' })) ||
-      (await click(page, { global: 'exactButton', text: 'Cancel' }));
-    if (!closed) await page.keyboard.press('Escape');
-    await wait(500);
-    return true;
-  }
-  return false;
 };
 
 const reportStepResult = ({ reporter, result, prefix, pageErrors }) => {
@@ -272,8 +257,8 @@ const reportStepResult = ({ reporter, result, prefix, pageErrors }) => {
   return [];
 };
 
-/** Run the scripted steps of a test. */
-const runSteps = async ({ page, pageErrors, steps, reporter }) => {
+/** Run the scripted manipulations of a test. */
+const runSteps = async ({ page, pageErrors, helper, steps, reporter }) => {
   const random = makeRandom(1);
   const failures = [];
   let performed = 0;
@@ -285,6 +270,7 @@ const runSteps = async ({ page, pageErrors, steps, reporter }) => {
     const result = await runStep({
       page,
       pageErrors,
+      helper,
       actionName,
       args,
       random,
@@ -318,13 +304,14 @@ const runSteps = async ({ page, pageErrors, steps, reporter }) => {
     failures.push('none of the drag and drops moved anything');
     reporter.log('   ❌ none of the drag and drops moved anything');
   }
-  return { failures, performed, skipped, drops, effectiveDrops };
+  return { failures, performed, skipped };
 };
 
 /** Run a random sequence of manipulations. */
 const runMonkey = async ({
   page,
   pageErrors,
+  helper,
   seed,
   steps,
   actionNames,
@@ -332,7 +319,7 @@ const runMonkey = async ({
   verbose,
 }) => {
   const random = makeRandom(seed);
-  const weights = require('./SpriteEditorActions').monkeyWeights;
+  const weights = helper.monkeyWeights;
   const weighted = [];
   (actionNames || Object.keys(weights)).forEach(name => {
     for (let index = 0; index < (weights[name] || 1); index++)
@@ -349,6 +336,7 @@ const runMonkey = async ({
     const result = await runStep({
       page,
       pageErrors,
+      helper,
       actionName,
       args: null,
       random,
@@ -376,11 +364,4 @@ const runMonkey = async ({
   return { failures, performed, skipped };
 };
 
-module.exports = {
-  makeRandom,
-  runSteps,
-  runMonkey,
-  runStep,
-  check,
-  closeAnyOverlay,
-};
+module.exports = { makeRandom, runStep, runSteps, runMonkey };
